Check type at setting Variable.grad

+++ chainer/functions/sigmoid_cross_entropy.py
@@ -42,7 +42,7 @@ class SigmoidCrossEntropy(function.Function):
 
     def backward_cpu(self, inputs, grad_outputs):
         t, gloss = inputs[1], grad_outputs[0]
-        gx = gloss * (self.y - t.astype(numpy.float32)) / t.shape[0]
+        gx = gloss * (self.y - t.astype(self.y.dtype)) / t.shape[0]
         return gx, None
 
     def backward_gpu(self, inputs, grad_outputs):
